requirements.txt
openai==1.88.0
python-dotenv==0.21.0
requests==2.31.0
aiohttp==3.10.8
pathlib==1.0.1
moviepy==1.0.3
pydub==0.25.1
pillow==9.5.0
scipy==1.12.0
opencv-python==4.9.0.80
assemblyai==0.20.2
srt_equalizer==0.1.8
whisperx==3.1.1
torch
torchaudio
tqdm==4.66.1
apscheduler==3.10.4
apprise==1.7.1
backoff==2.2.1
google-api-python-client==2.116.0
httplib2==0.22.0
oauth2client==4.1.3
pydantic==2.9.2
